Add the command definition basics.
A simple defcommand macro is available to define new macros. There is
only one level of commands supported right now.

(defcommand <command-name> (<args>)
  <docstring>
  <body>)

<command-name> can be any valid name (e.g. "create", "delete",
"pirate-arr"), and it will be what git-zed accepts.

The arguments are passed right away to the command. If the number of
arguments if invalid, the main function handles that properly by
displaying the usage.

The docstring is going to be used by the `usage` function to show the
doc for all the commands.